backend/ocr-nlp-service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
attrdict==2.0.1
Babel==2.14.0
bce-python-sdk==0.9.7
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.1.1
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.10.2
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fire==0.6.0
flask==3.0.3
flask-babel==4.0.0
Flask-Cors==4.0.0
fonttools==4.51.0
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
langcodes==3.4.0
language-data==1.2.0
lazy-loader==0.4
lmdb==1.4.1
lxml==5.2.1
marisa-trie==1.1.0
MarkupSafe==2.1.5
matplotlib==3.7.5
murmurhash==1.0.10
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.0.3
pdf2docx==0.5.8
pillow==10.3.0
premailer==3.10.0
preshed==3.0.9
protobuf==3.20.2
psutil==5.9.8
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic-core==2.18.2
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.8.1
rarfile==4.2
requests==2.31.0
ro-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ro_core_news_sm-3.7.0/ro_core_news_sm-3.7.0-py3-none-any.whl
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
termcolor==2.4.0
thinc==8.2.3
tifffile==2023.7.10
tqdm==4.66.2
typer==0.9.4
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
visualdl==2.5.3
wasabi==1.1.2
weasel==0.3.4
werkzeug==3.0.2
zipp==3.18.1
